modules/thermostat_reader/requirements.txt
paho-mqtt
azure-iot-device
python-dotenv